add TypeInfoFactoryExtractor
