+++ flink-core/src/main/java/org/apache/flink/api/java/typeutils/TypeExtractor.java
@@ -80,7 +80,6 @@ import static org.apache.flink.util.Preconditions.checkNotNull;
 import static org.apache.flink.api.java.typeutils.TypeExtractionUtils.checkAndExtractLambda;
 import static org.apache.flink.api.java.typeutils.TypeExtractionUtils.getAllDeclaredMethods;
 import static org.apache.flink.api.java.typeutils.TypeExtractionUtils.isClassType;
-import static org.apache.flink.api.java.typeutils.TypeExtractionUtils.sameTypeVars;
 import static org.apache.flink.api.java.typeutils.TypeExtractionUtils.typeToClass;
 
 /**
@@ -779,14 +778,14 @@ public class TypeExtractor {
 	private <IN1, IN2, OUT> TypeInformation<OUT> privateCreateTypeInfo(Class<?> baseClass, Class<?> clazz, int returnParamPos,
 			TypeInformation<IN1> in1Type, TypeInformation<IN2> in2Type) {
 
-		final List<Type> typeHierarchy = buildTypeHierarchy(clazz, baseClass);
+		final List<Type> typeHierarchy = TypeExtractionUtils.buildTypeHierarchy(clazz, baseClass);
 
 		if (typeHierarchy.size() < 1) {
 			throw new InvalidTypesException("The types of the interface " + baseClass.getName() + " could not be inferred. " +
 				"Support for synthetic interfaces, lambdas, and generic or raw types is limited at this point");
 		}
 
-		final Type baseClassType = resolveTypeFromTypeHierachy(typeHierarchy.get(typeHierarchy.size() - 1), typeHierarchy, false);
+		final Type baseClassType = TypeExtractionUtils.resolveTypeFromTypeHierachy(typeHierarchy.get(typeHierarchy.size() - 1), typeHierarchy, false);
 
 		final Type returnType = ((ParameterizedType) baseClassType).getActualTypeArguments()[returnParamPos];
 
@@ -818,11 +817,9 @@ public class TypeExtractor {
 
 		final List<TypeInformationExtractor<?>> typeInformationExtractors = findTypeInformationExtractors(t);
 
-		// check if type information can be created using a type factory
-//		final TypeInformation<OUT> typeFromFactory = createTypeInfoFromFactory(t, typeVariableBindings, currentExtractingClasses);
 		if (typeInformationExtractors.size() > 0) {
 			for (int i = 0; i < typeInformationExtractors.size(); i++) {
-				final TypeInformation<OUT> typeInformation = (TypeInformation<OUT>) typeInformationExtractors.get(i).create(t, new TypeInformationExtractorContext<>(typeVariableBindings, currentExtractingClasses));
+				final TypeInformation<OUT> typeInformation = (TypeInformation<OUT>) typeInformationExtractors.get(i).create(t, new TypeInformationExtractorContext(typeVariableBindings, currentExtractingClasses));
 				if (typeInformation != null) {
 					return typeInformation;
 				}
@@ -858,7 +855,7 @@ public class TypeExtractor {
 
 			typeHierarchy.add(curT);
 
-			curT = resolveTypeFromTypeHierachy(curT, typeHierarchy, true);
+			curT = TypeExtractionUtils.resolveTypeFromTypeHierachy(curT, typeHierarchy, true);
 
 			// create the type information for the subtypes
 			final TypeInformation<?>[] subTypesInfo = createSubTypesInfo(t, (ParameterizedType) curT, typeVariableBindings, extractingClasses, false);
@@ -1004,7 +1001,7 @@ public class TypeExtractor {
 		if (factory == null) {
 			return null;
 		}
-		final Type factoryDefiningType = resolveTypeFromTypeHierachy(factoryHierarchy.get(factoryHierarchy.size() - 1), factoryHierarchy, true);
+		final Type factoryDefiningType = TypeExtractionUtils.resolveTypeFromTypeHierachy(factoryHierarchy.get(factoryHierarchy.size() - 1), factoryHierarchy, true);
 
 		// infer possible type parameters from input
 		final Map<String, TypeInformation<?>> genericParams;
@@ -1210,7 +1207,7 @@ public class TypeExtractor {
 					throw new InvalidTypesException("Array type expected.");
 				}
 				if (component instanceof TypeVariable<?>) {
-					component = materializeTypeVariable(typeHierarchy, (TypeVariable<?>) component);
+					component = TypeExtractionUtils.materializeTypeVariable(typeHierarchy, (TypeVariable<?>) component);
 					if (component instanceof TypeVariable) {
 						return;
 					}
@@ -1229,7 +1226,7 @@ public class TypeExtractor {
 				}
 
 				if (component instanceof TypeVariable<?>) {
-					component = materializeTypeVariable(typeHierarchy, (TypeVariable<?>) component);
+					component = TypeExtractionUtils.materializeTypeVariable(typeHierarchy, (TypeVariable<?>) component);
 					if (component instanceof TypeVariable) {
 						return;
 					}
@@ -1254,7 +1251,7 @@ public class TypeExtractor {
 				}
 
 				if (component instanceof TypeVariable<?>) {
-					component = materializeTypeVariable(typeHierarchy, (TypeVariable<?>) component);
+					component = TypeExtractionUtils.materializeTypeVariable(typeHierarchy, (TypeVariable<?>) component);
 					if (component instanceof TypeVariable) {
 						return;
 					}
@@ -1309,7 +1306,7 @@ public class TypeExtractor {
 				validateIfWritable(typeInfo, type);
 			}
 		} else {
-			type = materializeTypeVariable(typeHierarchy, (TypeVariable<?>) type);
+			type = TypeExtractionUtils.materializeTypeVariable(typeHierarchy, (TypeVariable<?>) type);
 			if (!(type instanceof TypeVariable)) {
 				validateInfo(typeHierarchy, type, typeInfo);
 			}
@@ -1405,45 +1402,6 @@ public class TypeExtractor {
 			}
 		}
 		return fieldCount;
-	}
-
-	/**
-	 * Tries to find a concrete value (Class, ParameterizedType etc. ) for a TypeVariable by traversing the type hierarchy downwards.
-	 * If a value could not be found it will return the most bottom type variable in the hierarchy.
-	 */
-	private static Type materializeTypeVariable(List<Type> typeHierarchy, TypeVariable<?> typeVar) {
-		TypeVariable<?> inTypeTypeVar = typeVar;
-		// iterate thru hierarchy from top to bottom until type variable gets a class assigned
-		for (int i = typeHierarchy.size() - 1; i >= 0; i--) {
-			Type curT = typeHierarchy.get(i);
-
-			// parameterized type
-			if (curT instanceof ParameterizedType) {
-				Class<?> rawType = ((Class<?>) ((ParameterizedType) curT).getRawType());
-
-				for (int paramIndex = 0; paramIndex < rawType.getTypeParameters().length; paramIndex++) {
-
-					TypeVariable<?> curVarOfCurT = rawType.getTypeParameters()[paramIndex];
-
-					// check if variable names match
-					if (sameTypeVars(curVarOfCurT, inTypeTypeVar)) {
-						Type curVarType = ((ParameterizedType) curT).getActualTypeArguments()[paramIndex];
-
-						// another type variable level
-						if (curVarType instanceof TypeVariable<?>) {
-							inTypeTypeVar = (TypeVariable<?>) curVarType;
-						}
-						// class
-						else {
-							return curVarType;
-						}
-					}
-				}
-			}
-		}
-		// can not be materialized, most likely due to type erasure
-		// return the type variable of the deepest level
-		return inTypeTypeVar;
 	}
 
 	/**
@@ -1601,7 +1559,7 @@ public class TypeExtractor {
 			TypeVariable<?> fieldTypeGeneric = null;
 			if(fieldType instanceof TypeVariable) {
 				fieldTypeGeneric = (TypeVariable<?>) fieldType;
-				fieldType = materializeTypeVariable(typeHierarchy, (TypeVariable<?>)fieldType);
+				fieldType = TypeExtractionUtils.materializeTypeVariable(typeHierarchy, (TypeVariable<?>)fieldType);
 			}
 			for(Method m : clazz.getMethods()) {
 				final String methodNameLow = m.getName().endsWith("_$eq") ?
@@ -1683,7 +1641,7 @@ public class TypeExtractor {
 			try {
 				List<Type> fieldTypeHierarchy = new ArrayList<Type>(typeHierarchy);
 				fieldTypeHierarchy.add(fieldType);
-				Type resolveFeildType = resolveTypeFromTypeHierachy(fieldType, fieldTypeHierarchy, true);
+				Type resolveFeildType = TypeExtractionUtils.resolveTypeFromTypeHierachy(fieldType, fieldTypeHierarchy, true);
 
 				TypeInformation<?> ti = createTypeInfo(resolveFeildType, typeVariableBindings, extractingClasses);
 				pojoFields.add(new PojoField(field, ti));
@@ -1952,72 +1910,6 @@ public class TypeExtractor {
 	}
 
 	/**
-	 * Build the parameterized type hierarchy from subClass to baseClass.
-	 * @param subClass the begin class of the type hierarchy
-	 * @param baseClass the end class of the type hierarchy
-	 * @return the type hierarchy.
-	 */
-	static List<Type> buildTypeHierarchy(final Class<?> subClass, final Class<?> baseClass) {
-
-		final List<Type> typeHierarchy = new ArrayList<>();
-
-		if (baseClass.equals(subClass) || !baseClass.isAssignableFrom(subClass)) {
-			return Collections.emptyList();
-		}
-
-		final Type[] interfaceTypes = subClass.getGenericInterfaces();
-
-		for (Type type : interfaceTypes) {
-			if (baseClass.isAssignableFrom(typeToClass(type))) {
-				final List<Type> subTypeHierarchy = buildTypeHierarchy(typeToClass(type), baseClass);
-				if (type instanceof ParameterizedType) {
-					typeHierarchy.add(type);
-				}
-				typeHierarchy.addAll(subTypeHierarchy);
-				return typeHierarchy;
-			}
-		}
-
-		if (baseClass.isAssignableFrom(subClass)) {
-			final Type type = subClass.getGenericSuperclass();
-			final List<Type> subTypeHierarchy = buildTypeHierarchy(typeToClass(type), baseClass);
-			if (type instanceof ParameterizedType) {
-				typeHierarchy.add(type);
-			}
-			typeHierarchy.addAll(subTypeHierarchy);
-
-			return typeHierarchy;
-		}
-		return typeHierarchy;
-	}
-
-	/**
-	 * Resolve all {@link TypeVariable}s of the type from the type hierarchy.
-	 * @param type the type needed to be resovled
-	 * @param typeHierarchy the set of types which the {@link TypeVariable} could be resovled from.
-	 * @param resolveGenericArray whether to resolve the {@code GenericArrayType} or not. This is for compatible.
-	 *                               (Some code path resolves the component type of a GenericArrayType. Some code path
-	 *                               does not resolve the component type of a GenericArray. A example case is
-	 *                               {@link TypeExtractorTest#testParameterizedArrays()})
-	 * @return resolved type
-	 */
-	static Type resolveTypeFromTypeHierachy(final Type type, final List<Type> typeHierarchy, final boolean resolveGenericArray) {
-		Type resolvedType = type;
-
-		if (type instanceof TypeVariable) {
-			resolvedType = materializeTypeVariable(typeHierarchy, (TypeVariable) type);
-		}
-
-		if (resolvedType instanceof ParameterizedType) {
-			return resolveParameterizedType((ParameterizedType) resolvedType, typeHierarchy, resolveGenericArray);
-		} else if (resolveGenericArray && resolvedType instanceof GenericArrayType) {
-			return resolveGenericArrayType((GenericArrayType) resolvedType, typeHierarchy);
-		}
-
-		return resolvedType;
-	}
-
-	/**
 	 * Resolve all {@link TypeVariable}s of a {@link ParameterizedType}.
 	 * @param parameterizedType the {@link ParameterizedType} needed to be resovled.
 	 * @param typeHierarchy the set of types which the {@link TypeVariable}s could be resolved from.
@@ -2030,7 +1922,7 @@ public class TypeExtractor {
 
 		int i = 0;
 		for (Type type : parameterizedType.getActualTypeArguments()) {
-			actualTypeArguments[i] = resolveTypeFromTypeHierachy(type, typeHierarchy, resolveGenericArray);
+			actualTypeArguments[i] = TypeExtractionUtils.resolveTypeFromTypeHierachy(type, typeHierarchy, resolveGenericArray);
 			i++;
 		}
 
@@ -2048,7 +1940,7 @@ public class TypeExtractor {
 	 */
 	static Type resolveGenericArrayType(final GenericArrayType genericArrayType, final List<Type> typeHierarchy) {
 
-		final Type resolvedComponentType = resolveTypeFromTypeHierachy(genericArrayType.getGenericComponentType(), typeHierarchy, true);
+		final Type resolvedComponentType = TypeExtractionUtils.resolveTypeFromTypeHierachy(genericArrayType.getGenericComponentType(), typeHierarchy, true);
 
 		return new ResolvedGenericArrayType(genericArrayType.getTypeName(), resolvedComponentType);
 	}
@@ -2137,7 +2029,7 @@ public class TypeExtractor {
 			return typeVariableBindings;
 		}
 
-		final List<Type> functionTypeHierarchy = buildTypeHierarchy(clazz, baseClazz);
+		final List<Type> functionTypeHierarchy = TypeExtractionUtils.buildTypeHierarchy(clazz, baseClazz);
 
 		if (functionTypeHierarchy.size() < 1) {
 			return typeVariableBindings;
@@ -2147,13 +2039,13 @@ public class TypeExtractor {
 
 		if (in1TypeInfo != null) {
 			final Type in1Type = baseClass.getActualTypeArguments()[in1Pos];
-			final Type resolvedIn1Type = resolveTypeFromTypeHierachy(in1Type, functionTypeHierarchy, false);
+			final Type resolvedIn1Type = TypeExtractionUtils.resolveTypeFromTypeHierachy(in1Type, functionTypeHierarchy, false);
 			typeVariableBindings = bindTypeVariablesWithTypeInformationFromInput(resolvedIn1Type, in1TypeInfo);
 		}
 
 		if (in2TypeInfo != null) {
 			final Type in2Type = baseClass.getActualTypeArguments()[in2Pos];
-			final Type resolvedIn2Type = resolveTypeFromTypeHierachy(in2Type, functionTypeHierarchy, false);
+			final Type resolvedIn2Type = TypeExtractionUtils.resolveTypeFromTypeHierachy(in2Type, functionTypeHierarchy, false);
 			if (!typeVariableBindings.equals(Collections.emptyMap())) {
 				typeVariableBindings.putAll(bindTypeVariablesWithTypeInformationFromInput(resolvedIn2Type, in2TypeInfo));
 			}
@@ -2174,15 +2066,15 @@ public class TypeExtractor {
 
 		if (factory != null) {
 			final Type factoryDefiningType =
-				resolveTypeFromTypeHierachy(factoryHierarchy.get(factoryHierarchy.size() - 1), factoryHierarchy, true);
+				TypeExtractionUtils.resolveTypeFromTypeHierachy(factoryHierarchy.get(factoryHierarchy.size() - 1), factoryHierarchy, true);
 			return bindTypeVariableFromGenericParameters(factoryDefiningType, inTypeInfo);
 		} else if (inType instanceof GenericArrayType) {
 			return bindTypeVariableFromGenericArray(inType, inTypeInfo);
 		} else if (inTypeInfo instanceof GenericClassAware && isClassType(inType) && ((GenericClassAware) inTypeInfo).getGenericClass().isAssignableFrom(typeToClass(inType))) {
 			final List<Type> typeHierarchy = new ArrayList<Type>(Arrays.asList(inType));
 			final Class<?> genericClass = ((GenericClassAware) inTypeInfo).getGenericClass();
-			typeHierarchy.addAll(buildTypeHierarchy(typeToClass(inType), genericClass));
-			final Type resolvedTupleType = resolveTypeFromTypeHierachy(typeHierarchy.get(typeHierarchy.size() - 1), typeHierarchy, true);
+			typeHierarchy.addAll(TypeExtractionUtils.buildTypeHierarchy(typeToClass(inType), genericClass));
+			final Type resolvedTupleType = TypeExtractionUtils.resolveTypeFromTypeHierachy(typeHierarchy.get(typeHierarchy.size() - 1), typeHierarchy, true);
 
 			return bindTypeVariableFromGenericParameters(resolvedTupleType, inTypeInfo);
 		} else if (inTypeInfo instanceof PojoTypeInfo && isClassType(inType)) {
@@ -2254,7 +2146,7 @@ public class TypeExtractor {
 		List<Field> fields = getAllDeclaredFields(typeToClass(type), false);
 		for (Field field : fields) {
 			final Type fieldType = field.getGenericType();
-			final Type resolvedFieldType =  resolveTypeFromTypeHierachy(fieldType, pojoHierarchy, true);
+			final Type resolvedFieldType =  TypeExtractionUtils.resolveTypeFromTypeHierachy(fieldType, pojoHierarchy, true);
 			final Map sub = bindTypeVariablesWithTypeInformationFromInput(resolvedFieldType, getTypeOfPojoField(typeInformation, field));
 			typeVariableBindings.putAll(sub);
 		}
@@ -2265,18 +2157,7 @@ public class TypeExtractor {
 
 	private static final Map<Class<?>, TypeInformationExtractor<?>> REGISTERED_TYPEINFORMATION_EXTRACTOR = new HashMap<>();
 
-	interface TypeInformationExtractor<T> {
-
-		List<Class<?>> getTypeHandled();
-
-		TypeInformation<? extends T> create(Type type, Context<T> context);
-	}
-
-	interface Context<T> {
-		TypeInformation<T> create(Type type);
-	}
-
-	class TypeInformationExtractorContext<T> implements Context<T> {
+	class TypeInformationExtractorContext implements TypeInformationExtractor.Context {
 
 		final Map<TypeVariable<?>, TypeInformation<?>> typeVariableBindings;
 
@@ -2290,9 +2171,9 @@ public class TypeExtractor {
 		}
 
 		@Override
-		public TypeInformation<T> create(Type type) {
+		public TypeInformation<?> create(Type type) {
 			if (type instanceof TypeVariable) {
-				return (TypeInformation<T>) typeVariableBindings.get(type);
+				return typeVariableBindings.get(type);
 			}
 			return new TypeExtractor().createTypeInfo(type, typeVariableBindings, extractingClasses);
 		}
@@ -2343,68 +2224,5 @@ public class TypeExtractor {
 			result.addAll(findRegisteredTypeInformationExtractor(c));
 		}
 		return result;
-	}
-
-	private static class TypeInfoFactoryExtractor<T> implements TypeInformationExtractor<T> {
-
-		private final TypeInfoFactory<T> typeInfoFactory;
-
-		private final Type factoryType;
-
-		public TypeInfoFactoryExtractor(TypeInfoFactory<T> typeInfoFactory, Type factoryType) {
-			this.typeInfoFactory = typeInfoFactory;
-			this.factoryType = factoryType;
-		}
-
-		@Override
-		public List<Class<?>> getTypeHandled() {
-			throw new UnsupportedOperationException(String.format("%s does not support getTypeHandled.", getClass()));
-		}
-
-		@Override
-		public TypeInformation<? extends T> create(Type type, Context<T> context) {
-
-			final List<Type> factoryHierarchy = new ArrayList<>(Arrays.asList(type));
-			factoryHierarchy.addAll(buildTypeHierarchy(typeToClass(type), typeToClass(factoryType)));
-
-			final Type factoryDefiningType = resolveTypeFromTypeHierachy(factoryHierarchy.get(factoryHierarchy.size() - 1), factoryHierarchy, true);
-			final Map<String, TypeInformation<?>> genericParams;
-
-			if (factoryDefiningType instanceof ParameterizedType) {
-				genericParams = new HashMap<>();
-				final ParameterizedType paramDefiningType = (ParameterizedType) factoryDefiningType;
-				final Type[] args = typeToClass(paramDefiningType).getTypeParameters();
-
-				final TypeInformation<?>[] subtypeInfo = createSubTypesInfo(paramDefiningType, context);
-				assert subtypeInfo != null;
-				for (int i = 0; i < subtypeInfo.length; i++) {
-					genericParams.put(args[i].toString(), subtypeInfo[i]);
-				}
-			} else {
-				genericParams = Collections.emptyMap();
-			}
-
-			final TypeInformation<? extends T> createdTypeInfo = typeInfoFactory.createTypeInfo(type, genericParams);
-
-			if (createdTypeInfo == null) {
-				throw new InvalidTypesException("TypeInfoFactory returned invalid TypeInformation 'null'");
-			}
-			return createdTypeInfo;
-		}
-
-		private TypeInformation<?>[] createSubTypesInfo(final ParameterizedType definingType, final Context<T> context) {
-
-			final int typeArgumentsLength = definingType.getActualTypeArguments().length;
-			final TypeInformation<?>[] subTypesInfo = new TypeInformation<?>[typeArgumentsLength];
-			for (int i = 0; i < typeArgumentsLength; i++) {
-				final Type acutalTypeArgument = definingType.getActualTypeArguments()[i];
-				try {
-					subTypesInfo[i] = context.create(acutalTypeArgument);
-				} catch (InvalidTypesException e) {
-					subTypesInfo[i] = null;
-				}
-			}
-			return subTypesInfo;
-		}
 	}
 }
